packages/aiprod-cloud/pyproject.toml
[project]
name = "aiprod-cloud"
version = "0.1.0"
description = "AIPROD Cloud Integrations — Optional package for GCP, Stripe, S3, HuggingFace Hub"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }

dependencies = [
    "aiprod-core",
]

[project.optional-dependencies]
gcp = [
    "google-cloud-storage>=2.18.0",
    "google-cloud-logging>=3.11.0",
    "google-cloud-monitoring>=2.23.0",
    "google-api-core>=2.19.0",
]
gemini = [
    "google-generativeai>=0.8.0",
]
stripe = [
    "stripe>=10.0.0",
]
s3 = [
    "boto3>=1.35.0",
]
huggingface = [
    "huggingface-hub[hf-xet]>=0.31.4",
]
all = [
    "aiprod-cloud[gcp,gemini,stripe,s3,huggingface]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aiprod_cloud"]
